recommender/pipeline/components/training/requirements.txt
absl-py==2.1.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
alembic==1.14.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
astunparse==1.6.3 ; python_version >= "3.10" and python_full_version <= "3.12.2"
cachetools==5.5.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
certifi==2024.12.14 ; python_version >= "3.10" and python_full_version <= "3.12.2"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
colorama==0.4.6 ; python_version >= "3.10" and python_full_version <= "3.12.2" and (sys_platform == "win32" or platform_system == "Windows")
colorlog==6.9.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
db-dtypes==1.3.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
faiss-cpu==1.9.0.post1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
flatbuffers==24.12.23 ; python_version >= "3.10" and python_full_version <= "3.12.2"
gast==0.4.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-api-core==2.24.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-auth-oauthlib==0.4.6 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-auth==2.37.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-cloud-bigquery==3.14.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-crc32c==1.6.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-pasta==0.2.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
google-resumable-media==2.7.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
googleapis-common-protos==1.63.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
greenlet==3.1.1 ; python_full_version <= "3.12.2" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.10"
grpcio==1.69.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
h5py==3.12.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
idna==3.10 ; python_version >= "3.10" and python_full_version <= "3.12.2"
joblib==1.4.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
keras-tuner[bayesian]==1.4.7 ; python_version >= "3.10" and python_full_version <= "3.12.2"
keras==2.11.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
kt-legacy==1.0.5 ; python_version >= "3.10" and python_full_version <= "3.12.2"
libclang==18.1.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
mako==1.3.8 ; python_version >= "3.10" and python_full_version <= "3.12.2"
markdown==3.7 ; python_version >= "3.10" and python_full_version <= "3.12.2"
markupsafe==3.0.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
numpy==1.26.4 ; python_version >= "3.10" and python_full_version <= "3.12.2"
oauthlib==3.2.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
opt-einsum==3.4.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
optuna==4.2.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
packaging==24.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pandas==2.2.3 ; python_version >= "3.10" and python_full_version <= "3.12.2"
proto-plus==1.25.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
protobuf==3.19.6 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pyarrow==18.1.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pyasn1==0.6.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pytz==2024.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
pyyaml==6.0.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
requests==2.32.3 ; python_version >= "3.10" and python_full_version <= "3.12.2"
rsa==4.9 ; python_version >= "3.10" and python_full_version <= "3.12.2"
scikit-learn==1.6.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
scipy==1.15.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
setuptools==75.8.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
six==1.17.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
sqlalchemy==2.0.37 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorboard-data-server==0.6.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorboard-plugin-wit==1.8.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorboard==2.11.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorflow-estimator==2.11.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorflow-io-gcs-filesystem==0.37.1 ; python_version >= "3.10" and python_full_version <= "3.12.2" and platform_machine != "arm64" or python_version >= "3.10" and python_full_version <= "3.12.2" and platform_system != "Darwin"
tensorflow-ranking==0.5.5 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorflow-serving-api==2.11.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tensorflow==2.11.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
termcolor==2.5.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
threadpoolctl==3.5.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tqdm==4.67.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
tzdata==2024.2 ; python_version >= "3.10" and python_full_version <= "3.12.2"
urllib3==2.3.0 ; python_version >= "3.10" and python_full_version <= "3.12.2"
werkzeug==3.1.3 ; python_version >= "3.10" and python_full_version <= "3.12.2"
wheel==0.45.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
wrapt==1.17.1 ; python_version >= "3.10" and python_full_version <= "3.12.2"
xgboost==1.7.6 ; python_version >= "3.10" and python_full_version <= "3.12.2"
